move file+gophermap --> filecontents.go to indicate parent interface
